requirements.txt
pandas
flask
google-generativeai
python-dotenv
